Bugfixes & Building Particles
Added Factory Smoke and Damaged Smoke to Buildings.

Added checks to every CardEffect that ends a Building Damage effect if no valid buildings are on the board.

Updated Thief Cards with their Second Steal variable not referencing the Collected Pile.

Updated Regeneration point scoring a tad.

Switched PersistentCardManager foreach loop in DiscardPersistentCardProcess to a for loop.